+++ kai/kai_common.h
@@ -184,6 +184,14 @@ struct kai_rhs_pack_qsi8cx_params {
     float scale_multiplier;  ///< Product of input (refers to lhs and rhs) and output quantization scales.
 };
 
+/// Parameter struct for RHS matrix packing (Quantized Symmetric Integer 4-bit with per-block quantizatio and s1s0
+/// nibble ordering)
+struct kai_rhs_pack_nxk_qsi4c32p_qsu4c32s1s0_params {
+    int8_t lhs_zero_point;
+    uint8_t rhs_zero_point;
+    enum kai_datatype scale_dt;
+};
+
 /// Parameter struct for RHS matrix packing
 struct kai_rhs_pack_qs4cxs1s0_param {
     int8_t lhs_zero_point;   ///< LHS Matrix quantization zero-point